+++ iam.tf
@@ -10,3 +10,19 @@ resource "google_project_iam_member" "gke_sa_iam_member" {
   role    = var.iam_roles_list[count.index]
   member  = "serviceAccount:${google_service_account.gke_sa.email}"
 }
+
+
+#-------------------------------------
+# Workload Identity
+#-------------------------------------
+resource "google_service_account" "wi_gsa" {
+  account_id   = "basic-wi-gsa"
+  display_name = "Workload Identity Google service account"
+}
+
+resource "google_project_iam_member" "wi_gsa_iam_member" {
+  project = var.project_id
+  count   = length(var.wi_iam_roles_list)
+  role    = var.wi_iam_roles_list[count.index]
+  member  = "serviceAccount:${google_service_account.wi_gsa.email}"
+}